Separate styles for each form element
