+++ src/components/forms/checkboxField.css
@@ -1,0 +1,3 @@
+.tickInput {
+  margin-right: 0.5em;
+}
