+++ src/components/forms/fieldset.css
@@ -1,0 +1,4 @@
+.fieldset {
+  width: 500px;
+  margin-bottom: 2em;
+}
