+++ src/components/forms/form.css
@@ -1,0 +1,2 @@
+.form {
+}
