+++ src/components/forms/form.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import style from './forms.css';
+import style from './form.css';
 
 export default function Form(props) {
   return <form {...props} className={ style.form }/>;
