+++ src/components/forms/input.css
@@ -1,18 +1,3 @@
-.fieldset {
-  width: 500px;
-  margin-bottom: 2em;
-}
-
-.inputField {
-  margin-bottom: 0.5em;
-}
-
-.label {
-  cursor: pointer;
-  display: block;
-  margin-bottom: 4px;
-}
-
 .textInput {
   -webkit-appearance: textfield;
   border-radius: 2px;
@@ -32,8 +17,4 @@
 .textInput:focus {
   border-color: $warm-grey;
   outline: none;
-}
-
-.tickInput {
-  margin-right: 0.5em;
 }
