+++ src/components/forms/input.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import style from './forms.css';
+import style from './input.css';
 
 export default function Input(props) {
   return <input {...props} className={ style.textInput }/>;
