+++ src/components/forms/inputField.css
@@ -1,0 +1,3 @@
+.inputField {
+  margin-bottom: 0.5em;
+}
