+++ src/components/forms/inputField.js
@@ -3,7 +3,7 @@ import React, { PropTypes } from 'react';
 import Label from './label';
 import Input from './input';
 
-import styles from './forms.css';
+import styles from './inputField.css';
 
 export default function InputField(props) {
   let { id, label, type, ...other } = props;
@@ -11,7 +11,7 @@ export default function InputField(props) {
   type = type || 'text';
 
   return <div className={ styles.inputField }>
-    <Label htmlFor={ id } className={ styles.label }>{ label }:</Label>
+    <Label htmlFor={ id }>{ label }:</Label>
     <Input id={ id } type={ type } {...other} />
   </div>;
 }
