+++ src/components/forms/label.css
@@ -1,0 +1,5 @@
+.label {
+  cursor: pointer;
+  display: block;
+  margin-bottom: 4px;
+}
