+++ src/components/forms/label.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import style from './forms.css';
+import style from './label.css';
 
 export default function Label(props) {
   return <label {...props} className={ style.label }/>;
